Updates sample to work with latest version of SecureXPC
While this sample now works, integrating the latest version identified an issue with the SecureXPC framework: https://github.com/trilemma-dev/SecureXPC/issues/70

It means for right now when using the sample, uninstalling and reinstalling the helper tool (or updating it), the next action involving the helper tool will fail. The subsequent call and beyond will succeed.